Move decoder/validator to payloads/validation package
